requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
duckdb==0.8.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.5
fonttools==4.41.0
frozenlist==1.3.3
gunicorn==21.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
multidict==6.0.4
nltk==3.8.1
numpy==1.25.1
openai==0.27.6
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyarrow==12.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.6.0
python-magic==0.4.27
python-socketio==5.8.0
pytz==2023.3
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
threadpoolctl==3.2.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
yarl==1.9.2
